+++ src/main/java/com/netflix/backend/services/ProfileService.java
@@ -4,5 +4,5 @@ import com.netflix.backend.DTO.ProfileObject;
 
 public interface ProfileService {
     void addProfile(ProfileObject profile);
-    String deleteProfile(int profileId);
+    String deleteProfile(long profileId);
 }
